Only load required definitions
Parsing all the OpenAPI defintions takes quite some time. Changed the
definition loading to only parse definitions which are used by the CLI.

If there is no subcommand specified, it loads only the filenames of the
definitions. If there is a subcommand, it loads only that specific
definition. When autocomplete is requested, it checks the command
parameter and applies the same logic.