Avoid loading files in memory
Refactored the code to use readers and writers where possible in order to
avoid loading large files in memory. This allows the user to
upload/download multi GB files to storage.

Extended the file reference struct to take a file path. The file will
only be opened when the HTTP request is made in the executor. The file
will be directly streamed avoiding loading files in memory.