Upgrade dependendencies to latest major versions
Upgraded the following dependencies to the latest major versions:

urfave/cli/v2 --> urfave/cli/v3

The urfave/cli/v3 package comes with a couple of breaking changes.

The sub commands need to implement now the `CommandNotFound` function
so that missing sub commands are properly handled and show a detailed
error message.

Auto separting slice flags also needs to be disabled now on each sub
command and values need to be read through the `StringSlice` method.

EnvVar support needs to be enabled through a sources chain.

The package also provides now context support for actions. The context
needs to be propagated through all layers which will be done in a follow
up PR.

gopkg.in/yaml.v2 --> goccy/go-yaml

The gopkg.in/yaml package is officially unmaintained and deprecated.
Replaced it with the latest version of the goccy/go-yaml package.

This required changing the auth property map to have string keys.